requirements.txt
# Main dependencies for DiscordBot-Py-
discord.py>=2.0.0
python-dotenv
aiosqlite
Pillow
